// study-engine-ui/src/lib/generated/DueResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CardWithQuestion } from "./CardWithQuestion";

export type DueResponse = { cards: Array<CardWithQuestion>, dueCount: number, newCount: number, newRemaining: number, mode: string, };