// study-engine-ui/src/lib/generated/QuestionsResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CardWithQuestion } from "./CardWithQuestion";

export type QuestionsResponse = { cert: string, certName: string, domains: { [key in string]: string }, questions: Array<CardWithQuestion>, };